refactor(firefox): use arkenfox-userjs with a new method
Refactors the Firefox configuration to use `arkenfox-userjs` in a
different way. Previously, the `extraConfig` option was used to apply
the user preferences, which is no longer supported.

The new configuration uses `pkgs.firefox.override` to add
`arkenfox-userjs/user.cfg` and a local `overlay.js` file to
`extraPrefsFiles`. The `overlay.js` file is also updated to use
`lockPref` instead of `user_pref`.